requirements.txt
openai>=1.0.0,<3.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
